refactor: 팀 작성 폼 컴포넌트 분리
